requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
gpt4all-j>=0.1.0

